move FormatOptions to loci.formats
