+++ components/formats-api/src/loci/formats/in/DefaultMetadataOptions.java
@@ -32,6 +32,9 @@
 
 package loci.formats.in;
 
+import loci.formats.FormatOptions;
+
+
 /**
  * Default implementation of {@link loci.formats.in.MetadataOptions}.
  */
